refactor(ui-kit): use `img` element instead of mui `Avatar` for `AppLogo`
Mui avatar has some styling issues (such as default circular/square
shape) which makes it inappropriate for Rosen logos which have a title
and are not square shaped.

Use HTML `img` elements for rendering the logo and make it infer its
height based on the provided width and the image ratio.

In addition:
- remove `title` prop as it is not needed
- rename component to `AppLogo`