chore(swr-helpers): switch to a hard-coded baseURL
Based on the current state of our apps, they need to work in a docker
container. To do so, set `baseURL` to a hard-coded `'/api'` route and
let the docker container handle the redirection from the mentioned route
to accessible service routes.

Note that this is a temporary solution and need to be changed in favor
of a dynamic baseURL which can be set in the apps based on user config.