fix(watcher-app): remove bigint types from all api responses
In watcher app, it is supposed that number values have number type, but
because of the response transformation using `json-bigint` in `fetcher`
of `swr-helpers`, those number values get converted to bigint, which is
unexpected based on current apps implementation.

Remove the transformer temporarily and convert all bigint types to
number types in api responses. To fix the bug correctly, we need a
transformer that accepts the keys path that need to be transformed to
bigint.

Related issue:
local:ergo/rosen-bridge/ui#65