feat(rosen-app): add temporary build bash script
Vercel has a character limitation for entering build command and cannot
include all of monorepo package build scripts, so a bash script
containing build scripts of all monorepo packages (exactly the same as
ci build step) is required.

The bash script should be removed when a monorepo build tool is added to
the repo.