fix(rosen-app): fix wrong max transferable amount
Some native token is used for fee and min box value for the change box;
so when a user tries to transfer all of his/her native tokens, an
insufficient balance error is thrown.

Fix the issue by subtracting the required native assets for fee and min
box value from the maximum available balance.