refactor(ui-kit): switch from Vite to tsc
There is an issue with NextJS (or Vite, or Webpack): When NextJS
`transpilePackages` option is enabled, some errors are thrown which
prevents the ui from being rendered. It seems an issue with how esm/cjs
is interpreted in Webpack, but I'm not sure.

To fix the issue, opt out of Vite and switch to tsc for building the
package, applying all required changes (e.g. in tsconfig, etc.).

By doing so, another weird issue happens, which is described in
local:ergo/rosen-bridge/ui#193. In summary, `NoSsr` cannot be imported
from ui-kit anymore and needs to be imported directly from mui.

To address the latter issue, remove `NoSsr` from ui-kit and change all
`NoSsr` imports to mui.